Fixed error and reorganized file

+++ src/main/java/frc/acq/AcqCommand.java
@@ -3,5 +3,5 @@ package frc.acq;
 public abstract class AcqCommand {
 
     //Will be called by subsystem to execute this command
-    public abstract int execute();
+    public abstract double execute();
 }
